backend-v2/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.107.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.3
openai-agents>=0.3.3
openai-chatkit>=0.0.1
tavily-python>=0.3.0
